moved lib functions outside of plugins dir
